smartpc_v2/smartpc-advisor/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
mistralai>=1.0.0
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
